requirements.txt
flask
flask-cors
werkzeug
pdfplumber
python-dotenv
requests
sqlalchemy
psycopg2-binary